AMA-635-replace-vula-email-with-cilthelpdesk
update file and rename, move strings replace to config.py
